% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuumRemoved-package.R
\docType{package}
\name{ContinuumRemoved-package}
\alias{ContinuumRemoved}
\alias{ContinuumRemoved-package}
\title{ContinuumRemoved: Continuum Removal of Soil Spectral Reflectance Data}
\description{
The package calculates the continuum removal of a data set containing the spectral reflectance of post-dried soil samples collected from the Hudson Bay Lowlands. This normalizes the reflectance spectra, which allows the absorption features of individual soil samples to be compared from a common baseline.

The package calculates the continuum removal of a data set containing the spectral reflectance of post-dried soil samples collected from the Hudson Bay Lowlands. This normalizes the reflectance spectra, which allows the absorption features of individual soil samples to be compared from a common baseline.
}
\references{
\url{https://github.com/KangyuSo/my-GEO712-repository}
}
\author{
\strong{Maintainer}: Kangyu So \email{sok11@mcmaster.ca}


Kangyu So, School of Earth, Environment and Society, McMaster University \email{sok11@mcmaster.ca}
}
\keyword{internal}
