% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpred.R
\name{cpred}
\alias{cpred}
\title{Create a linear regression model, cross-validate it, and use it to predict the organic carbon content of soil samples based on their reflectance}
\source{
The data was collected by Kangyu So, Dr. Alemu Gonsamo, and Dr. Cheryl A. Rogers (All affiliated with McMaster University) from the Hudson Bay Lowlands region on July 2023
}
\usage{
cpred(data, reflectance, carbon, test_size, folds, prediction)
}
\arguments{
\item{data}{The dataset containing the average reflectance and organic carbon content of soil samples}

\item{reflectance}{The average reflectance at a given wavelength range}

\item{carbon}{The organic carbon content by percent dry weight}

\item{test_size}{The test size for the model}

\item{folds}{The number of folds for the cross-validation}

\item{prediction}{The name for the predicted organic carbon content output}
}
\value{
A column containing the predicted organic carbon content using a cross-validated linear regression model
}
\description{
Create a linear regression model, cross-validate it, and use it to predict the organic carbon content of soil samples based on their reflectance
}
\examples{
data <- data.frame(ref_dry_avg)
pred_c <- cpred(data, "Reflectance", "Organic_Carbon", 0.7, 10, "Predicted_Organic_Carbon")
}
\references{
https://github.com/KangyuSo/my-GEO712-repository
}
