% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbcontpred-package.R
\docType{data}
\name{pred_c}
\alias{pred_c}
\title{Reflectance values and measured and predicted organic carbon content for post-dried soil samples.}
\format{
A data frame with 303 observations and 4 variables:
\describe{
\item{Sample_ID}{ID of the soil sample}
\item{Reflectance}{Average spectral reflectance value for a given wavelength for the soil sample}
\item{Organic_Carbon}{Organic carbon content by percent dry weight for the soil sample}
\item{Predicted_Organic_Carbon}{Predicted organic carbon content by percent dry weight for the soil sample}
}
}
\usage{
data(pred_c)
}
\description{
A dataset containing information about the spectral reflectance, measured
organic carbon content, and predicted organic carbon content of 303
post-dried soil samples collected from the Hudson Bay Lowlands.
}
\examples{
 data(pred_c)
 Predicted_Organic_Carbon <- pred_c$Predicted_Organic_Carbon
}
\keyword{Soil}
\keyword{Terrestrial}
\keyword{carbon}
\keyword{content}
\keyword{datasets}
\keyword{estimation}
\keyword{modeling}
\keyword{spatial}
