% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuumRemoved-package.R
\docType{data}
\name{RefDryContinuumRemoved}
\alias{RefDryContinuumRemoved}
\title{Reflectance values for post-dried soil samples.}
\format{
A data frame with 2151 rows and 433 variables:
\describe{
\item{Wavelength}{Specific wavelength in nanometers}
\item{Sample ID}{ID of the soil sample}
\item{Reflectance}{Spectral reflectance value at a given wavelength}
}
}
\usage{
data(RefDryContinuumRemoved)
}
\description{
A dataset containing information about the spectral reflectance of 432
post-dried soil samples collected from the Hudson Bay Lowlands. The range of
information is between the wavelengths 350 nm and 2500 nm in increments of
1 nm.
}
\examples{
 data(RefDryContinuumRemoved)
 Wavelength <- RefDryContinuumRemoved$Wavelength
}
\keyword{Soil}
\keyword{carbon}
\keyword{datasets}
\keyword{spectroscopy}
\keyword{stock}
