% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbcontpred-package.R
\docType{package}
\name{carbcontpred-package}
\alias{carbcontpred}
\alias{carbcontpred-package}
\title{carbcontpred: Linear Regression of Soil Spectral Reflectance and Organic Carbcon Content Data}
\description{
The package builds and cross-validates a linear regression model for a dataset containing the spectral reflectance and organic carbon content of soil samples. The linear regression model uses the reflectance to predict the organic carbon content.

This package is an exercise in package creation using R studio. The package
includes a function that builds and implements a linear regression model and
a sample dataset containing the reflectance and organic carbon data of soil
samples. The package also includes the documentation for the function and
dataset.
}
\references{
\url{https://github.com/KangyuSo/my-GEO712-repository}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KangyuSo/carbcontpred}
  \item Report bugs at \url{https://github.com/KangyuSo/carbcontpred/issues}
}

}
\author{
\strong{Maintainer}: Kangyu So \email{sok11@mcmaster.ca}


Kangyu So, School of Earth, Environment and Society, McMaster University \email{sok11@mcmaster.ca}
}
\keyword{internal}
