% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappred.R
\name{mappred}
\alias{mappred}
\title{The function first extracts an algebraic equation based on the reflectance and predicted organic carbon content of soil samples. Then the function applies continuum removal over the reflectance values of an Region of Interest (ROI) data table collected from image processing and analysis software ENVI. Then the function applies the extracted algebraic equation on this attribute table to predict organic carbon content of points based on their reflectance values.}
\source{
The data used for the predictions was collected by Kangyu So, Dr. Alemu Gonsamo, and Dr. Cheryl A. Rogers (All affiliated with McMaster University) from the Hudson Bay Lowlands region on July 2022. The ROI dataset was collected from images collected by the Moderate Resolution Imaging Spectroradiometer (MODIS) satellite operated by the National Aeronautics and Space Administration (NASA).
}
\usage{
mappred(predictions_dataframe, ROI_dataset_path, output_dataset)
}
\arguments{
\item{predictions_dataframe}{The dataset resulting from the output of the cpred function, contains the average reflectance and measured and predicted organic carbon content of some soil samples}

\item{ROI_dataset_path}{The ROI data table exported from ENVI into a CSV file, contains geoinformation and reflectance values of the ROI divided into individual points}

\item{output_dataset}{The updated ROI data table containing the new predicted organic carbon content of the points}
}
\value{
A CSV file containing the ROI data table as well as the predicted organic carbon content of the points using an algebraic equation extracted from the predictions of a cross-validated linear regression model
}
\description{
The function first extracts an algebraic equation based on the reflectance and predicted organic carbon content of soil samples. Then the function applies continuum removal over the reflectance values of an Region of Interest (ROI) data table collected from image processing and analysis software ENVI. Then the function applies the extracted algebraic equation on this attribute table to predict organic carbon content of points based on their reflectance values.
}
\examples{
mappred_c <- mappred(pred_c, "data-raw/Reflectance_ROI.csv", "ROI_Predicted_Carbon")
}
\references{
https://github.com/KangyuSo/my-GEO712-repository
}
